+++ integration-test-java/src/test/java/org/seasar/doma/it/criteria/EntityqlBatchInsertTest.java
@@ -1,6 +1,7 @@
 package org.seasar.doma.it.criteria;
 
 import static java.util.stream.Collectors.toList;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -9,18 +10,22 @@ import java.util.Collections;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.seasar.doma.it.Dbms;
 import org.seasar.doma.it.IntegrationTestEnvironment;
+import org.seasar.doma.it.Run;
 import org.seasar.doma.jdbc.BatchResult;
 import org.seasar.doma.jdbc.Config;
 import org.seasar.doma.jdbc.criteria.Entityql;
+import org.seasar.doma.jdbc.dialect.Dialect;
 
 @ExtendWith(IntegrationTestEnvironment.class)
 public class EntityqlBatchInsertTest {
-
   private final Entityql entityql;
+  private final Dialect dialect;
 
   public EntityqlBatchInsertTest(Config config) {
     this.entityql = new Entityql(config);
+    this.dialect = config.getDialect();
   }
 
   @Test
@@ -61,5 +66,81 @@ public class EntityqlBatchInsertTest {
 
     BatchResult<Employee> result = entityql.insert(e, Collections.emptyList()).execute();
     assertTrue(result.getEntities().isEmpty());
+  }
+
+  @Test
+  @Run(onlyIf = {Dbms.MYSQL, Dbms.POSTGRESQL}) // TODO: Implement it to work in other dialects
+  public void batch_onDuplicateKeyUpdate() {
+    Department_ d = new Department_();
+
+    Department department1 = new Department();
+    department1.setDepartmentId(5);
+    department1.setDepartmentNo(50);
+    department1.setDepartmentName("PLANNING");
+    department1.setLocation("TOKYO");
+    Department department2 = new Department();
+    department2.setDepartmentId(1);
+    department2.setDepartmentNo(60);
+    department2.setDepartmentName("DEVELOPMENT");
+    department2.setLocation("KYOTO");
+
+    List<Department> list = Arrays.asList(department1, department2);
+
+    BatchResult<Department> result = entityql.insert(d, list).onDuplicateKeyUpdate().execute();
+    if (dialect.getName().equals("mysql") || dialect.getName().equals("mariadb")) {
+      assertArrayEquals(new int[] {1, 2}, result.getCounts());
+    } else {
+      assertArrayEquals(new int[] {1, 1}, result.getCounts());
+    }
+    assertEquals(list, result.component1());
+
+    Department resultDepartment1 =
+        entityql.from(d).where(c -> c.eq(d.departmentId, department1.getDepartmentId())).fetchOne();
+    Department resultDepartment2 =
+        entityql.from(d).where(c -> c.eq(d.departmentId, department2.getDepartmentId())).fetchOne();
+    // inserted
+    assertEquals(50, resultDepartment1.getDepartmentNo());
+    assertEquals("PLANNING", resultDepartment1.getDepartmentName());
+    assertEquals("TOKYO", resultDepartment1.getLocation());
+    // updated
+    assertEquals(60, resultDepartment2.getDepartmentNo());
+    assertEquals("DEVELOPMENT", resultDepartment2.getDepartmentName());
+    assertEquals("KYOTO", resultDepartment2.getLocation());
+  }
+
+  @Test
+  @Run(onlyIf = {Dbms.MYSQL, Dbms.POSTGRESQL}) // TODO: Implement it to work in other dialects
+  public void batch_onDuplicateKeyIgnore() {
+    Department_ d = new Department_();
+
+    Department department1 = new Department();
+    department1.setDepartmentId(5);
+    department1.setDepartmentNo(50);
+    department1.setDepartmentName("PLANNING");
+    department1.setLocation("TOKYO");
+    Department department2 = new Department();
+    department2.setDepartmentId(1);
+    department2.setDepartmentNo(60);
+    department2.setDepartmentName("DEVELOPMENT");
+    department2.setLocation("KYOTO");
+
+    List<Department> list = Arrays.asList(department1, department2);
+
+    BatchResult<Department> result = entityql.insert(d, list).onDuplicateKeyIgnore().execute();
+    assertArrayEquals(new int[] {1, 0}, result.getCounts());
+    assertEquals(list, result.component1());
+
+    Department resultDepartment1 =
+        entityql.from(d).where(c -> c.eq(d.departmentId, department1.getDepartmentId())).fetchOne();
+    Department resultDepartment2 =
+        entityql.from(d).where(c -> c.eq(d.departmentId, department2.getDepartmentId())).fetchOne();
+    // inserted
+    assertEquals(50, resultDepartment1.getDepartmentNo());
+    assertEquals("PLANNING", resultDepartment1.getDepartmentName());
+    assertEquals("TOKYO", resultDepartment1.getLocation());
+    // ignored
+    assertEquals(10, resultDepartment2.getDepartmentNo());
+    assertEquals("ACCOUNTING", resultDepartment2.getDepartmentName());
+    assertEquals("NEW YORK", resultDepartment2.getLocation());
   }
 }
